Move CA1008 to its new location. There's some codecleanup involved here.
