+++ PortFxCop/src/Analyzer.Utilities/Extensions/IEnumerableExtensions.cs
@@ -7,7 +7,7 @@ using System.Linq;
 
 namespace Analyzer.Utilities
 {
-    internal static class IEnumerableExtensions
+    public static class IEnumerableExtensions
     {
         public static IEnumerable<T> Concat<T>(this IEnumerable<T> source, T value)
         {
