+++ PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/CSharp/CSharpEnumsShouldHaveZeroValue.Fixer.cs
@@ -1,20 +1,16 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-                                 
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+
 using System.Composition;
-using System.Diagnostics;  
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CodeFixes;
-using Microsoft.CodeAnalysis.CSharp;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
 
 namespace Microsoft.ApiDesignGuidelines.Analyzers
-{                                 
+{
     /// <summary>
     /// CA1008: Enums should have zero value
     /// </summary>
     [ExportCodeFixProvider(LanguageNames.CSharp), Shared]
     public class CSharpEnumsShouldHaveZeroValueFixer : EnumsShouldHaveZeroValueFixer
-    { 
-
+    {
     }
 }
