+++ PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/Core/EnumsShouldHaveZeroValue.Fixer.cs
@@ -1,36 +1,166 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using System;
 using System.Collections.Generic;
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CodeFixes;
-using Microsoft.CodeAnalysis.Editing;
 using System.Collections.Immutable;
-using System.Composition;
+using System.Diagnostics;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Analyzer.Utilities;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CodeFixes;
+using Microsoft.CodeAnalysis.Editing;
 
 namespace Microsoft.ApiDesignGuidelines.Analyzers
-{                              
+{
     /// <summary>
     /// CA1008: Enums should have zero value
     /// </summary>
     public abstract class EnumsShouldHaveZeroValueFixer : CodeFixProvider
     {
-        public sealed override ImmutableArray<string> FixableDiagnosticIds => ImmutableArray.Create(EnumsShouldHaveZeroValueAnalyzer.RuleId);
-
-        public sealed override FixAllProvider GetFixAllProvider()
+        public sealed override ImmutableArray<string> FixableDiagnosticIds
         {
-            // See https://github.com/dotnet/roslyn/blob/master/docs/analyzers/FixAllProvider.md for more information on Fix All Providers
-            return WellKnownFixAllProviders.BatchFixer;
+            get { return ImmutableArray.Create(EnumsShouldHaveZeroValueAnalyzer.RuleId); }
         }
 
-        public sealed override async Task RegisterCodeFixesAsync(CodeFixContext context)
-        {                              
-            // This is to get rid of warning CS1998, please remove when implementing this analyzer
-            await new Task(() => { });
-            throw new NotImplementedException();
+        public override async Task RegisterCodeFixesAsync(CodeFixContext context)
+        {
+            var model = await context.Document.GetSemanticModelAsync(context.CancellationToken).ConfigureAwait(false);
+
+            var flagsAttributeType = WellKnownTypes.FlagsAttribute(model.Compilation);
+            if (flagsAttributeType == null)
+            {
+                return;
+            }
+
+            // We cannot have multiple overlapping diagnostics of this id.
+            var diagnostic = context.Diagnostics.Single();
+            var root = await context.Document.GetSyntaxRootAsync(context.CancellationToken).ConfigureAwait(false);
+            var node = root.FindNode(context.Span);
+
+            ISymbol declaredSymbol = model.GetDeclaredSymbol(node, context.CancellationToken);
+            Debug.Assert(declaredSymbol != null);
+
+            foreach (var customTag in diagnostic.Descriptor.CustomTags)
+            {
+                switch (customTag)
+                {
+                    case EnumsShouldHaveZeroValueAnalyzer.RuleRenameCustomTag:
+                        context.RegisterCodeFix(new DocumentChangeAction(MicrosoftApiDesignGuidelinesAnalyzersResources.EnumsShouldZeroValueFlagsRenameCodeFix,
+                                                    async ct => await GetUpdatedDocumentForRuleNameRenameAsync(context.Document, (IFieldSymbol)declaredSymbol, context.CancellationToken).ConfigureAwait(false)),
+                                                diagnostic);
+                        return;
+                    case EnumsShouldHaveZeroValueAnalyzer.RuleMultipleZeroCustomTag:
+                        context.RegisterCodeFix(new DocumentChangeAction(MicrosoftApiDesignGuidelinesAnalyzersResources.EnumsShouldZeroValueFlagsMultipleZeroCodeFix,
+                            async ct => await ApplyRuleNameMultipleZeroAsync(context.Document, (INamedTypeSymbol)declaredSymbol, context.CancellationToken).ConfigureAwait(false)),
+                        diagnostic);
+                        return;
+
+                    case EnumsShouldHaveZeroValueAnalyzer.RuleNoZeroCustomTag:
+                        context.RegisterCodeFix(new DocumentChangeAction(MicrosoftApiDesignGuidelinesAnalyzersResources.EnumsShouldZeroValueNotFlagsNoZeroValueCodeFix,
+                                                    async ct => await ApplyRuleNameNoZeroValueAsync(context.Document, (INamedTypeSymbol)declaredSymbol, context.CancellationToken).ConfigureAwait(false)),
+                                                diagnostic);
+                        return;
+                }
+            }
+        }
+        
+        private static SyntaxNode GetDeclaration(ISymbol symbol)
+        {
+            return (symbol.DeclaringSyntaxReferences.Length > 0) ? symbol.DeclaringSyntaxReferences[0].GetSyntax() : null;
+        }
+
+        private SyntaxNode GetExplicitlyAssignedField(IFieldSymbol originalField, SyntaxNode declaration, SyntaxGenerator generator)
+        {
+            var originalInitializer = generator.GetExpression(declaration);
+            if (originalInitializer != null || !originalField.HasConstantValue)
+            {
+                return declaration;
+            }
+
+            return generator.WithExpression(declaration, generator.LiteralExpression(originalField.ConstantValue));
+        }
+
+        private async Task<Document> GetUpdatedDocumentForRuleNameRenameAsync(Document document, IFieldSymbol field, CancellationToken cancellationToken)
+        {
+            var newSolution = await CodeAnalysis.Rename.Renamer.RenameSymbolAsync(document.Project.Solution, field, "None", null, cancellationToken).ConfigureAwait(false);
+            return newSolution.GetDocument(document.Id);
+        }
+
+        private async Task<Document> ApplyRuleNameMultipleZeroAsync(Document document, INamedTypeSymbol enumType, CancellationToken cancellationToken)
+        {
+            // Diagnostic: Remove all members that have the value zero from '{0}' except for one member that is named 'None'.
+            // Fix: Remove all members that have the value zero except for one member that is named 'None'.
+            var editor = SymbolEditor.Create(document);
+
+            bool needsNewZeroValuedNoneField = true;
+            var set = EnumsShouldHaveZeroValueAnalyzer.GetZeroValuedFields(enumType).ToSet();
+
+            bool makeNextFieldExplicit = false;
+            foreach (IFieldSymbol field in enumType.GetMembers().Where(m => m.Kind == SymbolKind.Field))
+            {
+                var isZeroValued = set.Contains(field);
+                var isZeroValuedNamedNone = isZeroValued && EnumsShouldHaveZeroValueAnalyzer.IsMemberNamedNone(field);
+
+                if (!isZeroValued || isZeroValuedNamedNone)
+                {
+                    if (makeNextFieldExplicit)
+                    {
+                        await editor.EditOneDeclarationAsync(field, (e, d) => e.ReplaceNode(d, GetExplicitlyAssignedField(field, d, e.Generator)), cancellationToken).ConfigureAwait(false);
+                        makeNextFieldExplicit = false;
+                    }
+
+                    if (isZeroValuedNamedNone)
+                    {
+                        needsNewZeroValuedNoneField = false;
+                    }
+                }
+                else
+                {
+                    await editor.EditOneDeclarationAsync(field, (e, d) => e.RemoveNode(d), cancellationToken).ConfigureAwait(false); // removes the field declaration
+                    makeNextFieldExplicit = true;
+                }
+            }
+
+            if (needsNewZeroValuedNoneField)
+            {
+                await editor.EditOneDeclarationAsync(enumType, (e, d) => e.InsertMembers(d, 0, new[] { e.Generator.EnumMember("None") }), cancellationToken).ConfigureAwait(false);
+            }
+
+            return editor.GetChangedDocuments().First();
+        }
+
+        private async Task<Document> ApplyRuleNameNoZeroValueAsync(Document document, INamedTypeSymbol enumType, CancellationToken cancellationToken)
+        {
+            var editor = SymbolEditor.Create(document);
+
+            // remove any non-zero member named 'None'
+            foreach (IFieldSymbol field in enumType.GetMembers().Where(m => m.Kind == SymbolKind.Field))
+            {
+                if (EnumsShouldHaveZeroValueAnalyzer.IsMemberNamedNone(field))
+                {
+                    await editor.EditOneDeclarationAsync(field, (e, d) => e.RemoveNode(d), cancellationToken).ConfigureAwait(false);
+                }
+            }
+
+            // insert zero-valued member 'None' to top
+            await editor.EditOneDeclarationAsync(enumType, (e, d) => e.InsertMembers(d, 0, new[] { e.Generator.EnumMember("None") }), cancellationToken).ConfigureAwait(false);
+
+            return editor.GetChangedDocuments().First();
+        }
+
+        protected virtual SyntaxNode GetParentNodeOrSelfToFix(SyntaxNode nodeToFix)
+        {
+            return nodeToFix;
+        }
+
+        private Document GetUpdatedDocumentWithFix(Document document, SyntaxNode root, SyntaxNode nodeToFix, IList<SyntaxNode> newFields, CancellationToken cancellationToken)
+        {
+            nodeToFix = GetParentNodeOrSelfToFix(nodeToFix);
+            var g = SyntaxGenerator.GetGenerator(document);
+            var newEnumSyntax = g.AddMembers(nodeToFix, newFields);
+            var newRoot = root.ReplaceNode(nodeToFix, newEnumSyntax);
+            return document.WithSyntaxRoot(newRoot);
         }
     }
 }
