+++ PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/Core/Microsoft.ApiDesignGuidelines.Analyzers.props
@@ -9,6 +9,7 @@
    <PropertyGroup>
       <CodeAnalysisRuleSetOverrides>
         $(CodeAnalysisRuleSetOverrides);
+        -Microsoft.Design#CA1008;
         -Microsoft.Design#CA1017;
         -Microsoft.Design#CA1018;
         -Microsoft.Design#CA1036;
